feat(web): new information architecture with consolidated nav, routes, and redirects
- One nav source of truth (components/layout/nav-items.ts) consumed by both
  the desktop sidebar and mobile nav; repo nav slims to 6 groups (Overview,
  Docs, Architecture, Code Health, People & History, Chat) with Settings and
  Usage & savings pinned last; Search lives in the sidebar and opens the
  command palette
- Overview becomes the repo landing page (/repos/[id] redirects); Chat moves
  to /repos/[id]/chat with seeded-question support via ?q=
- /architecture hosts the former /graph, /c4 and /symbols pages as ?view=
  sub-views (c4 internal view param renamed to c4view); old routes redirect
  with query passthrough
- /code-health mounts the former Health overview as the Triage tab plus the
  Risk tabs (heatmap/hotspots/modules/dead-code/impact/security); health
  sub-pages move to /code-health/{coverage,refactoring-targets,trend};
  /health/* and /risk redirect with tab mapping
- Zombie pages (/hotspots /ownership /dead-code /blast-radius /security)
  replaced with redirects into the new sections; orphan loading screens
  removed
- Internal links across ui and web point at the new routes; entity hrefs for
  owner/commit now target /owners/[id] and /commits
- Breadcrumb label map covers every segment; command palette gains per-repo
  page navigation and an Ask repowise quick-ask action
