+++ packages/web/src/app/repos/[id]/architecture/loading.tsx
@@ -1,5 +1,5 @@
 import { OwlLoader } from "@repowise-dev/ui/shared/owl-loader";
 
 export default function Loading() {
-  return <OwlLoader />;
+  return <OwlLoader label="Loading architecture…" />;
 }
