--- packages/web/src/app/repos/[id]/dead-code/loading.tsx
@@ -1,5 +1,0 @@
-import { OwlLoader } from "@repowise-dev/ui/shared/owl-loader";
-
-export default function DeadCodeLoading() {
-  return <OwlLoader />;
-}
