--- packages/web/src/app/repos/[id]/graph/loading.tsx
@@ -1,5 +1,0 @@
-import { OwlLoader } from "@repowise-dev/ui/shared/owl-loader";
-
-export default function GraphLoading() {
-  return <OwlLoader />;
-}
