--- packages/web/src/app/repos/[id]/ownership/loading.tsx
@@ -1,5 +1,0 @@
-import { OwlLoader } from "@repowise-dev/ui/shared/owl-loader";
-
-export default function OwnershipLoading() {
-  return <OwlLoader />;
-}
