+++ packages/web/src/app/repos/[id]/security/page.tsx
@@ -1,17 +1,11 @@
-"use client";
+import { redirect } from "next/navigation";
 
-import { useEffect } from "react";
-import { useParams, useRouter } from "next/navigation";
-
-// Security now lives as a tab under Risk. Keep this route as a redirect so
-// existing links / bookmarks continue to work.
-export default function SecurityPage() {
-  const params = useParams<{ id: string }>();
-  const router = useRouter();
-
-  useEffect(() => {
-    router.replace(`/repos/${params.id}/risk?tab=security`);
-  }, [params.id, router]);
-
-  return null;
+/** Legacy route — redirects into the consolidated IA. */
+export default async function LegacyRedirect({
+  params,
+}: {
+  params: Promise<{ id: string }>;
+}) {
+  const { id } = await params;
+  redirect(`/repos/${id}/code-health?tab=security`);
 }
