--- packages/web/src/app/repos/[id]/symbols/loading.tsx
@@ -1,5 +1,0 @@
-import { OwlLoader } from "@repowise-dev/ui/shared/owl-loader";
-
-export default function SymbolsLoading() {
-  return <OwlLoader />;
-}
