+++ packages/web/src/components/chat/chat-interface.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { useEffect, useRef } from "react";
 import { ChatInterface as ChatInterfaceShell } from "@repowise-dev/ui/chat/chat-interface";
 import { useChat } from "@/lib/hooks/use-chat";
 import { ModelSelector } from "./model-selector";
@@ -8,9 +9,11 @@ import { ConversationHistory } from "./conversation-history";
 interface ChatInterfaceProps {
   repoId: string;
   repoName?: string;
+  /** Question to send immediately on mount (quick-ask deep links, `?q=`). */
+  initialQuestion?: string;
 }
 
-export function ChatInterface({ repoId, repoName }: ChatInterfaceProps) {
+export function ChatInterface({ repoId, repoName, initialQuestion }: ChatInterfaceProps) {
   const {
     messages,
     conversationId,
@@ -20,6 +23,15 @@ export function ChatInterface({ repoId, repoName }: ChatInterfaceProps) {
     loadConversation,
     reset,
   } = useChat(repoId);
+
+  // Fire the seeded question exactly once (guards StrictMode double-effects).
+  const seededRef = useRef(false);
+  useEffect(() => {
+    if (!initialQuestion || seededRef.current) return;
+    seededRef.current = true;
+    void sendMessage(initialQuestion);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [initialQuestion]);
 
   return (
     <ChatInterfaceShell
